Move StarRatings to it's own component
